server/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
dataclasses
python-dotenv
pytest
httpx